// types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      liked_props: {
        Row: {
          americanOdds: number | null
          category: string | null
          created_at: string
          eventDate: string | null
          eventId: number | null
          eventName: string | null
          id: string
          label: string | null
          leagueId: number | null
          leagueName: string | null
          line: string | null
          propLabel: string | null
          providerOutcomeId: string | null
          userId: string | null
        }
        Insert: {
          americanOdds?: number | null
          category?: string | null
          created_at?: string
          eventDate?: string | null
          eventId?: number | null
          eventName?: string | null
          id?: string
          label?: string | null
          leagueId?: number | null
          leagueName?: string | null
          line?: string | null
          propLabel?: string | null
          providerOutcomeId?: string | null
          userId?: string | null
        }
        Update: {
          americanOdds?: number | null
          category?: string | null
          created_at?: string
          eventDate?: string | null
          eventId?: number | null
          eventName?: string | null
          id?: string
          label?: string | null
          leagueId?: number | null
          leagueName?: string | null
          line?: string | null
          propLabel?: string | null
          providerOutcomeId?: string | null
          userId?: string | null
        }
        Relationships: []
      }
      upcoming_fight_odds: {
        Row: {
          created_at: string
          f1_book_odds: number | null
          f1_pic_url: string | null
          f2_book_odds: number | null
          f2_pic_url: string | null
          fight_date: string | null
          fight_id: string
          fight_name: string | null
          fighter1: string | null
          fighter2: string | null
          id: number
          odds1: number | null
          odds2: number | null
        }
        Insert: {
          created_at?: string
          f1_book_odds?: number | null
          f1_pic_url?: string | null
          f2_book_odds?: number | null
          f2_pic_url?: string | null
          fight_date?: string | null
          fight_id: string
          fight_name?: string | null
          fighter1?: string | null
          fighter2?: string | null
          id?: number
          odds1?: number | null
          odds2?: number | null
        }
        Update: {
          created_at?: string
          f1_book_odds?: number | null
          f1_pic_url?: string | null
          f2_book_odds?: number | null
          f2_pic_url?: string | null
          fight_date?: string | null
          fight_id?: string
          fight_name?: string | null
          fighter1?: string | null
          fighter2?: string | null
          id?: number
          odds1?: number | null
          odds2?: number | null
        }
        Relationships: []
      }
      upcoming_nfl_odds: {
        Row: {
          created_at: string
          game_date: string | null
          game_id: string
          game_name: string | null
          odds1: number | null
          odds2: number | null
          opp: string | null
          opp_book_odds: number | null
          opp_name: string | null
          opp_pic_url: string | null
          team: string | null
          team_book_odds: number | null
          team_name: string | null
          team_pic_url: string | null
        }
        Insert: {
          created_at?: string
          game_date?: string | null
          game_id: string
          game_name?: string | null
          odds1?: number | null
          odds2?: number | null
          opp?: string | null
          opp_book_odds?: number | null
          opp_name?: string | null
          opp_pic_url?: string | null
          team?: string | null
          team_book_odds?: number | null
          team_name?: string | null
          team_pic_url?: string | null
        }
        Update: {
          created_at?: string
          game_date?: string | null
          game_id?: string
          game_name?: string | null
          odds1?: number | null
          odds2?: number | null
          opp?: string | null
          opp_book_odds?: number | null
          opp_name?: string | null
          opp_pic_url?: string | null
          team?: string | null
          team_book_odds?: number | null
          team_name?: string | null
          team_pic_url?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      delete_user: {
        Args: Record<PropertyKey, never>
        Returns: undefined
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}